refactor: rename DownloadDialog to SaveDialog to prepare for saving recording to account UX flow, extract download zip flow to utils
